+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - a total SPECIAL points display to the character sheet
 - a total Skill points display to the character sheet
+- a finalize data step for items
 
 ### Changed